requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.0
sqlparse==0.5.1
tzdata==2024.1
django-cors-headers==4.3.1